+++ tests/testthat/test-super.R
@@ -48,7 +48,7 @@ test_that("super() works with S3 objects", {
 })
 
 test_that("super() works with abstract S3 classes (#686)", {
-  gen <- new_generic("gen", "x")
+  gen := new_generic("x")
   method(gen, class_POSIXct) <- function(x) "POSIXct"
   method(gen, class_POSIXt) <- function(x) "POSIXt"
 
@@ -58,8 +58,9 @@ test_that("super() works with abstract S3 classes (#686)", {
 })
 
 test_that("super() works with S4 objects", {
-  Foo1 := local_S4_class(representation(x = "numeric"))
-  Foo2 := local_S4_class(contains = "Foo1")
+  local_S4_classes()
+  setClass("Foo1", representation(x = "numeric"))
+  setClass("Foo2", contains = "Foo1")
   obj <- methods::new("Foo2", x = 5)
 
   gen := new_generic("x")